requirements.txt
openai
fastapi==0.112.0
uvicorn==0.30.3
python-dotenv==1.0.1
langchain
langchain-community
langchain-openai
faiss-cpu==1.8.0
tiktoken
fastapi
uvicorn
pydantic
python-dotenv
numpy<2
cmake
pytest
python-multipart